requirements.txt
moviepy>=1.0.3
edge-tts>=6.1.9
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0
streamlit>=1.30.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
mutagen>=1.47.0
